refactor: menuGroup과 TableGroup을 다른 패키지로 분리
