+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -12,9 +12,9 @@ import java.util.stream.Collectors;
 import kitchenpos.menu.application.dto.MenuDto;
 import kitchenpos.menu.application.dto.MenuProductDto;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.domain.exception.MenuException;
+import kitchenpos.menu_group.domain.MenuGroupRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.vo.Price;
