+++ src/main/java/kitchenpos/table_group/application/TableGroupUnGroupValidationEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.application;
+package kitchenpos.table_group.application;
 
 import java.util.List;
