+++ src/main/java/kitchenpos/table_group/domain/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.table_group.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
