+++ src/test/java/kitchenpos/menu_group/application/MenuGroupServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu_group.application;
 
 import static kitchenpos.support.fixture.MenuFixture.순살파닭두마리메뉴_DTO;
 import static kitchenpos.support.fixture.MenuFixture.한마리메뉴_DTO;
@@ -7,7 +7,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import kitchenpos.menu.application.dto.MenuGroupDto;
+import kitchenpos.menu_group.application.MenuGroupDto;
 import kitchenpos.support.ServiceIntegrationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
