+++ src/test/java/kitchenpos/support/ServiceIntegrationTest.java
@@ -11,20 +11,20 @@ import static kitchenpos.support.fixture.TableFixture.비어있지_않는_주문
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu_group.application.MenuGroupService;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.application.dto.MenuDto;
-import kitchenpos.menu.application.dto.MenuGroupDto;
+import kitchenpos.menu_group.application.MenuGroupDto;
 import kitchenpos.menu.application.dto.MenuProductDto;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.application.dto.OrderDto;
 import kitchenpos.order.application.dto.OrderLineItemDto;
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.application.dto.ProductDto;
-import kitchenpos.table.application.TableGroupService;
+import kitchenpos.table_group.application.TableGroupService;
 import kitchenpos.table.application.TableService;
 import kitchenpos.table.application.dto.OrderTableDto;
-import kitchenpos.table.application.dto.TableGroupDto;
+import kitchenpos.table_group.application.dto.TableGroupDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
