+++ src/test/java/kitchenpos/table_group/application/TableGroupServiceTest.java
@@ -1,13 +1,13 @@
-package kitchenpos.table.application;
+package kitchenpos.table_group.application;
 
 import static java.util.stream.Collectors.toUnmodifiableList;
 import static kitchenpos.order.domain.exception.OrderExceptionType.ORDER_IS_NOT_COMPLETION;
 import static kitchenpos.support.fixture.TableFixture.비어있는_전쳬_주문_테이블_DTO;
 import static kitchenpos.support.fixture.TableFixture.비어있는_주문_테이블;
 import static kitchenpos.support.fixture.TableFixture.비어있지_않는_전쳬_주문_테이블_DTO;
-import static kitchenpos.table.domain.exception.TableGroupExceptionType.ORDER_TABLE_IS_NOT_EMPTY_OR_ALREADY_GROUPED;
-import static kitchenpos.table.domain.exception.TableGroupExceptionType.ORDER_TABLE_IS_NOT_PRESENT_ALL;
-import static kitchenpos.table.domain.exception.TableGroupExceptionType.ORDER_TABLE_SIZE_IS_LOWER_THAN_ZERO_OR_EMPTY;
+import static kitchenpos.table_group.domain.exception.TableGroupExceptionType.ORDER_TABLE_IS_NOT_EMPTY_OR_ALREADY_GROUPED;
+import static kitchenpos.table_group.domain.exception.TableGroupExceptionType.ORDER_TABLE_IS_NOT_PRESENT_ALL;
+import static kitchenpos.table_group.domain.exception.TableGroupExceptionType.ORDER_TABLE_SIZE_IS_LOWER_THAN_ZERO_OR_EMPTY;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.tuple;
@@ -24,11 +24,13 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.exception.OrderException;
 import kitchenpos.support.ServiceIntegrationTest;
+import kitchenpos.table.application.TableService;
 import kitchenpos.table.application.dto.OrderTableDto;
-import kitchenpos.table.application.dto.TableGroupDto;
+import kitchenpos.table_group.application.dto.TableGroupDto;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.table.domain.exception.TableGroupException;
+import kitchenpos.table_group.domain.exception.TableGroupException;
+import kitchenpos.table_group.application.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
